services/reviews_service/requirements.txt
fastapi==0.95.0
uvicorn[standard]==0.22.0
SQLAlchemy==2.0.21
psycopg2-binary==2.9.6
pydantic==1.10.11
python-jose==3.3.0
bleach==6.0.0
pytest==7.4.0
httpx==0.24.1
python-dotenv==1.0.0